# emergent_matter_model/openapi.yaml
openapi: 3.0.3
info:
  title: Emergent Matter Simulation API
  version: 1.2.0
  description: |
    REST API for emergent matter simulation. Entropy S(X,t) is a thermodynamic state
    field (the arrow of change per Second Law), not a spatial coordinate. The last
    entry in X_grid is the entropy state sweep; output M[..., j] is matter at state S_j.
servers:
  - url: http://127.0.0.1:5000
paths:
...
  /api/v1/verification/physics:
    get:
      summary: Run foundational physics known-limit verification
      operationId: verifyPhysics
      responses:
        '200':
          description: All known-limit checks passed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PhysicsVerificationReport'}
        '500':
          description: One or more scientific verification checks failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PhysicsVerificationReport'}
  /api/v1/verification/inference:
    get:
      summary: Run synthetic inference recovery and interval calibration
      operationId: verifyInference
      responses:
        '200':
          description: All injection-recovery and calibration checks passed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InferenceValidationReport'}
        '500':
          description: One or more inference validation checks failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InferenceValidationReport'}
  /api/v1/verification/engines:
    get:
      summary: Run cross-engine deterministic and synthetic verification
      operationId: verifyEngines
      responses:
        '200':
          description: All engine checks passed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EngineValidationReport'}
        '500':
          description: One or more engine checks failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EngineValidationReport'}
  /api/v1/predictions/commit:
    post:
      summary: Commit a prediction digest without retaining prediction content
      description: Disabled unless EMRF_API_ALLOW_PREDICTION_WRITE=1.
      operationId: commitPrediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, prediction]
              properties:
                label: {type: string, minLength: 1, maxLength: 100}
                prediction: {}
      responses:
        '201':
          description: Immutable commitment metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionCommitment'}
        '400': {description: Invalid prediction or label}
        '403': {description: Prediction writes disabled}
  /api/v1/predictions/{commitment_id}:
    get:
      summary: Retrieve prediction commitment metadata
      operationId: getPredictionCommitment
      parameters:
        - {name: commitment_id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Commitment metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PredictionCommitment'}
        '400': {description: Invalid commitment id}
        '404': {description: Unknown commitment}
  /api/v1/predictions/{commitment_id}/verify:
    post:
      summary: Verify revealed JSON against a prediction commitment
      operationId: verifyPrediction
      parameters:
        - {name: commitment_id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prediction]
              properties:
                prediction: {}
      responses:
        '200':
          description: Revealed prediction matches
        '409':
          description: Revealed prediction does not match
        '400': {description: Invalid request}
        '404': {description: Unknown commitment}
  /api/v1/commands:
    get:
      summary: List every registered program (same as `emrf list`)
      operationId: listCommands
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [data, analysis, stress-test, figures, visualization, service, audit, release]
      responses:
        '200':
          description: Registered programs
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_enabled: {type: boolean}
                  commands:
                    type: array
                    items: {$ref: '#/components/schemas/Command'}
        '400':
          description: Unknown category
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /api/v1/runs:
    get:
      summary: List durable scientific run records
      operationId: listRuns
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 50}}
      responses:
        '200':
          description: Recent runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs: {type: array, items: {$ref: '#/components/schemas/Run'}}
  /api/v1/runs/{run_id}:
    get:
      summary: Retrieve one durable run record
      operationId: getRun
      parameters:
        - {name: run_id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Run record, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        '404': {description: Unknown run}
        '409': {description: Corrupt run record}
  /api/v1/runs/{run_id}/resume:
    post:
      summary: Resume a failed, timed-out or interrupted registered command
      operationId: resumeRun
      parameters:
        - {name: run_id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                timeout: {type: number, exclusiveMinimum: 0, maximum: 600}
      responses:
        '200': {description: Resumed command finished}
        '400': {description: Run is not resumable}
        '403': {description: Command execution disabled}
        '404': {description: Unknown run}
        '504': {description: Resumed command timed out}
  /api/v1/jobs:
    get:
      summary: List durable queued jobs
      operationId: listJobs
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 50}}
      responses:
        '200':
          description: Recent jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: {type: array, items: {$ref: '#/components/schemas/Job'}}
    post:
      summary: Submit a registered command to the durable queue
      operationId: submitJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command: {type: string}
                args: {type: array, items: {type: string}}
                timeout: {type: number, exclusiveMinimum: 0}
                evidence_class:
                  type: string
                  enum: [observational, published, synthetic, illustrative, software]
                max_attempts: {type: integer, minimum: 1, maximum: 100, default: 1}
      responses:
        '202': {description: Job accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '400': {description: Invalid or non-runnable command}
        '403': {description: Command execution disabled}
        '404': {description: Unknown command}
  /api/v1/jobs/{job_id}:
    get:
      summary: Retrieve one queued job
      operationId: getJob
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Job record, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '404': {description: Unknown job}
  /api/v1/jobs/{job_id}/cancel:
    post:
      summary: Cancel a queued job
      operationId: cancelJob
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Cancelled job, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '400': {description: Job is not queued}
        '403': {description: Command execution disabled}
        '404': {description: Unknown job}
  /api/v1/datasets:
    get:
      summary: Curated observational data catalog (same as `emrf data catalog`)
      operationId: listDatasets
      parameters:
        - name: verify
          in: query
          required: false
          description: Re-hash managed objects (read-only)
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: Catalog datasets and managed-store status
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items: {$ref: '#/components/schemas/Dataset'}
...
  /api/v1/data/backups:
    post:
      summary: Create a verified portable data-library backup
      operationId: createDataBackup
      responses:
        '201': {description: Backup created}
        '403': {description: Data mutation disabled}
  /api/v1/data/backups/offsite:
    post:
      summary: Create a verified backup under an independent configured root
      operationId: createOffsiteDataBackup
      responses:
        '201': {description: Off-site backup created and verified}
        '400': {description: No independent destination configured}
        '403': {description: Data mutation disabled}
...
  /api/v1/data/restore-drills:
    post:
      summary: Run a full temporary backup and restore verification drill
      operationId: runDataRestoreDrill
      responses:
        '200': {description: Drill passed and durable receipt written}
        '403': {description: Data mutation disabled}
        '409': {description: Drill integrity failure}
  /api/v1/data/gc:
    post:
      summary: Find or remove unreferenced content-addressed objects
      operationId: garbageCollectData
      responses:
        '200': {description: Garbage-collection report}
        '403': {description: Destructive execution disabled}
...
  /api/v1/simulate:
    post:
      summary: Run emergent matter simulation
      operationId: simulateV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
            examples:
              basic4d:
                summary: 3 spatial dimensions + entropy state sweep
                value:
                  n: 4
                  weights: [0.3, 0.3, 0.2, 0.2]
                  X_grid:
                    - [-2, -1, 0, 1, 2]
                    - [-2, -1, 0, 1, 2]
                    - [-2, -1, 0, 1, 2]
                    - [0, 0.25, 0.5, 0.75, 1.0]
                  k: 1.0
                  alpha: 1.0
                  C0: 1.0
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /simulate:
    post:
      summary: Legacy simulation endpoint
      description: Backward-compatible alias for /api/v1/simulate.
      operationId: simulateLegacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Command:
      type: object
      properties:
        name: {type: string}
        script: {type: string}
        category: {type: string}
        summary: {type: string}
        network: {type: boolean}
        gui: {type: boolean}
        exists: {type: boolean}
...
    PhysicsVerificationReport:
      type: object
      required: [evidence_class, constants_reference, all_passed, checks]
      properties:
        evidence_class: {type: string, enum: [software, mixed]}
        constants_reference: {type: string}
        all_passed: {type: boolean}
        checks:
          type: array
          items:
            type: object
            required: [name, measured, expected, unit, relative_error, relative_tolerance, passed, reference, evidence_class]
            properties:
              name: {type: string}
              measured: {type: number}
              expected: {type: number}
              unit: {type: string}
              relative_error: {type: number}
              relative_tolerance: {type: number}
              passed: {type: boolean}
              reference: {type: string}
              evidence_class:
                type: string
                enum: [software, illustrative]
              limitation: {type: string, nullable: true}
    InferenceValidationReport:
      type: object
      required: [evidence_class, all_passed, realizations, random_seed, checks, truths, limitations]
      properties:
        evidence_class: {type: string, enum: [synthetic]}
        all_passed: {type: boolean}
        realizations: {type: integer, minimum: 16}
        random_seed: {type: integer, minimum: 0}
        truths:
          type: object
          additionalProperties: {type: number}
        limitations:
          type: array
          items: {type: string}
        checks:
          type: array
          items:
            type: object
            required: [name, metric, measured, expected, lower_bound, upper_bound, passed, reference, evidence_class, limitation]
            properties:
              name: {type: string}
              metric: {type: string}
              measured: {type: number}
              expected: {type: number}
              lower_bound: {type: number}
              upper_bound: {type: number}
              passed: {type: boolean}
              reference: {type: string}
              evidence_class: {type: string, enum: [synthetic]}
              limitation: {type: string}
    EngineValidationReport:
      type: object
      required: [evidence_class, all_passed, checks, limitations]
      properties:
        evidence_class: {type: string, enum: [mixed]}
        all_passed: {type: boolean}
        limitations: {type: array, items: {type: string}}
        checks:
          type: array
          items:
            type: object
            required: [name, metric, measured, expected, lower_bound, upper_bound, passed, reference, evidence_class]
            properties:
              name: {type: string}
              metric: {type: string}
              measured: {type: number}
              expected: {type: number}
              lower_bound: {type: number}
              upper_bound: {type: number}
              passed: {type: boolean}
              reference: {type: string}
              evidence_class: {type: string, enum: [software, synthetic, illustrative]}
              limitation: {type: string, nullable: true}
    PredictionCommitment:
      type: object
      required: [commitment_id, label, created_utc, sha256, bytes, media_type, evidence_class, content_stored]
      properties:
        commitment_id: {type: string}
        label: {type: string}
        created_utc: {type: string, format: date-time}
        sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        bytes: {type: integer, minimum: 1}
        source_name: {type: string, nullable: true}
        media_type: {type: string}
        evidence_class: {type: string, enum: [preregistered-prediction]}
        content_stored: {type: boolean, enum: [false]}
    Run:
      type: object
      properties:
        run_id: {type: string}
        command: {type: string}
        args: {type: array, items: {type: string}}
        evidence_class: {type: string}
        status: {type: string, enum: [pending, running, succeeded, failed, timed_out, interrupted]}
        attempt: {type: integer}
        created_utc: {type: string, format: date-time}
        started_utc: {type: string, format: date-time, nullable: true}
        completed_utc: {type: string, format: date-time, nullable: true}
        returncode: {type: integer, nullable: true}
        duration_seconds: {type: number, nullable: true}
        environment_sha256: {type: string}
        git_commit: {type: string, nullable: true}
        git_dirty: {type: boolean, nullable: true}
        inputs: {type: object, additionalProperties: {type: string}}
        seeds: {type: object, additionalProperties: {type: integer}}
...